requirements.txt
Flask
PyYAML
redis